requirements.txt
beautifulsoup4==4.6.0
requests==2.21.0
pymongo==3.6.1
daiquiri==1.5.0
schedule==0.5.0
tweepy==3.7.0
python-pixabay==2.0
praw==5.3.0
pyimgur==0.6.0
vimeo==0.1.9
dailymotion==0.2.2
flickrapi==2.4.0
newsapi==0.1.1
google-api-python-client==1.7.6
google-auth==1.6.1
google-auth-oauthlib==0.2.0
google-auth-httplib2==0.0.3
